set stracture
