+++ src/blocks.js
@@ -36,7 +36,9 @@ import './blocks/testimonials/block.js';
 import './blocks/advanced-gallery/block.js';
 import './blocks/form/block.js';
 import './blocks/table-of-contents/block.js';
-import './blocks/food/block.js';
+import './blocks/restaurant-menu/block.js';
+import './blocks/restaurant-menu/menu-category/block.js';
+import './blocks/restaurant-menu/menu-category/category-item/block.js';
 
 import './plugins/kadence-control.js';
 import './plugins/editor-width.js';
