+++ src/blocks/restaurant-menu/attributes.js
@@ -1,0 +1,3 @@
+export default {
+
+}
