+++ src/blocks/restaurant-menu/editor.scss
@@ -1,0 +1,290 @@
+/**
+ * #.# Editor Styles
+ *
+ * CSS for just Backend enqueued after style.scss
+ * which makes it higher in priority.
+ */
+ .components-panel__row button.components-button.kb-add-field {
+    margin: 0 auto;
+    min-width: 60%;
+    text-align: center;
+    margin-bottom: 10px;
+    margin-top: -10px;
+    align-items: center;
+    justify-content: center;
+}
+.components-button-group.kb-radio-button-flex-fill {
+    display: flex;
+    width: 100%;
+}
+.components-button-group.kb-radio-button-flex-fill button {
+    flex: 1;
+    text-align: center;
+    justify-content: center;
+}
+.components-panel__row button.components-button.kb-add-field svg {
+    width: 16px;
+    height: 16px;
+    margin-top: 2px;
+    margin-right: 4px;
+}
+.edit-post-sidebar button.kb-add-option svg {
+    width: 16px;
+    height: 16px;
+    margin-top: 2px;
+    margin-right: 4px;
+}
+.edit-post-sidebar button.kb-add-option {
+     -ms-flex-align: center;
+    align-items: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+    margin-bottom:24px
+}
+.edit-post-settings-sidebar__panel-block .components-panel__body .components-base-control.kb-option-text-control {
+    margin-bottom: 10px;
+}
+.kb-form .kadence-blocks-form-field .kb-forms-submit.is-selected {
+	min-width: 70px;
+}
+.field-options-wrap {
+    position: relative;
+}
+.field-options-wrap .kadence-blocks-list-item__control-menu {
+    top:0;
+}
+// Controls
+.kb-form-align-center .kb-form {
+    justify-content: center;
+}
+.kb-form-align-right .kb-form {
+    justify-content: flex-end;
+}
+.kb-form {
+	display: flex;
+	flex-wrap: wrap;
+	margin: 0 -5px;
+	align-items: flex-end;
+	.kadence-blocks-form-field {
+		box-sizing: border-box;
+		padding: 0 5px;
+		margin-bottom: 16px;
+		flex-wrap: wrap;
+		align-items: center;
+		display: flex;
+		width: 100%;
+		&.kb-submit-field {
+			margin-bottom: 0;
+		}
+		.kb-placeholder {
+			opacity: 0.5;
+		}
+		label {
+			display: block;
+			a {
+				pointer-events: none;
+			}
+		}
+		.kb-checkbox-style {
+			padding: 0;
+			background-color: initial;
+			cursor: default;
+			-webkit-appearance: checkbox;
+			appearance: checkbox;
+			box-sizing: border-box;
+			margin: 3px 5px 3px 0;
+			border-color: #818a91;
+			box-shadow: none;
+			&:before {
+				display:none;
+			}
+		}
+		.kb-select-style-field {
+			color: #373a3c;
+			border: 1px solid #818a91;
+			border-radius: 3px;
+			padding: 10px 12px;
+			min-height: 40px;
+			flex-basis: 100%;
+			max-width: 100%;
+			width: 100%;
+			flex-grow: 1;
+			vertical-align: middle;
+			//-webkit-appearance: select;
+			-webkit-appearance: menulist;
+			appearance: menulist;
+			background:initial;
+			// background-image: url('data:image/svg+xml;charset=US-ASCII,%3Csvg%20width%3D%2220%22%20height%3D%2220%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%3E%3Cpath%20d%3D%22M5%206l5%205%205-5%202%201-7%207-7-7%202-1z%22%20fill%3D%22%23555%22%2F%3E%3C%2Fsvg%3E');
+			// background-position: right 5px top 55%;
+			// background-repeat: no-repeat;
+			// background-size: 16px 16px;
+			cursor: pointer;
+		}
+		&.kb-input-size-small {
+			.kb-select-style-field {
+				min-height: 28px;
+			}
+		}
+		&.kb-input-size-large {
+			.kb-select-style-field {
+				min-height: 60px;
+			}
+		}
+		.kb-select-style-field[multiple] {
+			height: auto;
+			max-height: 70px;
+			-webkit-appearance: listbox;
+			appearance: listbox;
+		}
+		.kb-text-style-field {
+			flex-basis: 100%;
+			max-width: 100%;
+			width: 100%;
+			border: 1px solid #818a91;
+			background-color: #fff;
+			color: #373a3c;
+			vertical-align: middle;
+			flex-grow: 1;
+			z-index: 10;
+			min-height: 40px;
+			padding: 10px 12px;
+			border-radius: 3px;
+			outline:0;
+			&:focus {
+				outline:0;
+			}
+		}
+		.kb-checkbox-style-field, .kb-radio-style-field {
+			flex-basis: 100%;
+			max-width: 100%;
+			width: 100%;
+			flex-grow: 1;
+			margin-top: 0.3em;
+		}
+		.kb-checkbox-item, .kb-radio-item {
+			display: flex;
+			margin: 0;
+			align-items: center;
+		}
+		&.kb-input-size-custom {
+			.kb-text-style-field {
+				min-height: 0;
+				padding:10px;
+			}
+		}
+		&.kb-input-size-small {
+			.kb-text-style-field {
+				min-height: 28px;
+				padding: 5px 10px;
+			}
+		}
+		&.kb-input-size-large {
+			.kb-text-style-field {
+				min-height: 60px;
+				padding: 16px 16px;
+			}
+		}
+		.kb-forms-submit {
+			z-index: 1;
+			padding: 8px 16px;
+			cursor: text;
+			position: relative;
+			font-size: 18px;
+			border-width:0px;
+			border-radius: 3px;
+			border-color:#1768ea;
+			background:#1768ea;
+			color:#ffffff;
+			display: flex;
+			text-align: center;
+			justify-content: center;
+			border-style: solid;
+			overflow: hidden;
+			transition: all .3s ease-in-out;
+			&.kb-button-size-small {
+				padding: 4px 8px;;
+				font-size: 16px;
+			}
+			&.kb-button-size-large {
+				padding: 12px 24px;
+				font-size: 20px;
+			}
+			&.kb-button-width-full {
+				width: 100%;
+			}
+		}
+		.kb-forms-submit::before {
+			position: absolute;
+			content: "";
+			top: 0;
+			right: 0;
+			bottom: 0;
+			left: 0;
+			z-index: -1;
+			opacity: 0;
+			transition: all .3s ease-in-out;
+		}
+		.kb-forms-submit:hover::before {
+			opacity: 1;
+		}
+		.kb-field-help {
+			color: #777;
+			margin: 2px 0 5px 0;
+			font-size: 80%;
+			font-style: italic;
+			display: block;
+			flex-basis: 100%;
+			width: 0;
+		}
+		a.kb-accept-link {
+			display: block;
+			flex-basis: 100%;
+			width: 0;
+			pointer-events: none;
+		}
+	}
+	.kadence-blocks-form-field.is-selected {
+		padding-top: 24px;
+		position:relative;
+		&:before {
+			content: '';
+			position: absolute;
+			border: 4px solid #0085ba;
+			left: 0;
+			right: 0;
+			top: 0;
+			bottom: -4px;
+			z-index: 1;
+		}
+		&:after {
+			content: '';
+			left: 0;
+			right: 0;
+			position: absolute;
+			top: 0;
+			height: 24px;
+			background: #0085ba;
+		}
+	}
+	.kadence-blocks-field-item-controls {
+		display:flex;
+		position:absolute;
+		top:2px;
+		left:2px;
+		z-index: 10;
+		background: #0085ba;
+		.components-button {
+			padding: 0;
+			height:auto;
+			color:#f4f4f4;
+		}
+		&.kadence-blocks-field-item__inline-menu {
+			left:auto;
+			right:2px;
+		}
+	}
+}
+.input-size-large input {
+	min-height: 60px;
+    padding: 10px 16px;
+}
