+++ src/blocks/restaurant-menu/menu-category/attributes.js
@@ -1,0 +1,8 @@
+const { __ } = wp.i18n;
+
+export default {
+	menuTitle: {
+		type: 'String',
+		default: __( 'APPETIZERS' )
+ 	}
+}
