+++ src/blocks/restaurant-menu/menu-category/category-item/attributes.js
@@ -1,0 +1,18 @@
+export default {
+	title: {
+		type: 'String'
+	},
+
+	description: {
+		type: 'String'
+	},
+
+	currency: {
+		type: 'String',
+		default: '$'
+	},
+
+	price: {
+		type: 'String'
+	}
+}
