+++ src/blocks/restaurant-menu/menu-category/category-item/icons.js
@@ -1,0 +1,17 @@
+/**
+ * WordPress dependencies
+ */
+const { Path, SVG, G } = wp.components;
+
+const icons = {}
+
+icons.block =
+	<SVG width="18px" height="18px" viewBox="0 0 18 18" version="1.1" xmlns="http://www.w3.org/2000/svg">
+	    <G id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+	        <G id="restaurant-cutlery-symbol-of-a-cross" fill="#189CD4" fillRule="nonzero">
+	            <Path d="M17.4416,16.4584 C17.7228,16.7144 17.8196,17.3412 17.4564,17.7012 C17.0956,18.0648 16.4676,17.9684 16.2128,17.6876 C13.8556,15.1564 11.4996,12.624 9.1424,10.092 L8.8744,9.806 C6.4088,12.4532 3.944,15.1016 1.4792,17.75 C1.2252,18.0304 0.5984,18.1272 0.2376,17.7636 C-0.1264,17.4032 -0.0304,16.7748 0.2504,16.52 C2.9072,14.0488 5.562,11.5768 8.2184,9.1044 C8.2184,9.1044 6.9328,7.7176 6.3784,7.1224 C5.0812,8.0524 3.1588,7.6988 2.0032,6.272 C0.7604,4.7668 -0.3116,2.0748 0.726,0.9724 C1.8296,-0.066 4.52,1.0084 6.0252,2.25 C7.4516,3.4068 7.8072,5.3264 6.8752,6.6244 C7.5432,7.2468 8.2112,7.8676 8.88,8.4896 C8.8784,8.4908 8.8776,8.4916 8.876,8.4928 C8.876,8.4928 8.9176,8.5256 8.88,8.4896 C9.5048,7.9076 10.13,7.326 10.7552,6.744 C10.2272,5.8424 10.578,4.578 11.4356,3.8564 C12.4192,3.032 16.0636,-0.0968 16.1616,0.0024 C16.2592,0.0996 13.7032,2.5148 11.9956,4.2276 L12.2804,4.512 C14.1032,2.9144 16.6764,0.5184 16.7756,0.616 C16.8732,0.714 14.3972,3.208 12.7452,4.978 L13.028,5.26 C14.7952,3.6068 17.29,1.1304 17.3868,1.2276 C17.4848,1.326 15.0884,3.8996 13.4908,5.7224 L13.7736,6.0048 C15.4856,4.2972 17.8996,1.7412 17.9984,1.838 C18.0972,1.9364 14.1512,6.56 14.1456,6.5668 C14.1448,6.566 14.9832,5.582 14.1424,6.5644 C13.34,7.5032 12.1416,7.7872 11.2556,7.2464 C10.6804,7.8636 10.1064,8.4808 9.532,9.098 C9.4816,9.0504 14.8056,14.0052 17.4416,16.4584 Z" id="Path"></Path>
+	        </G>
+	    </G>
+	</SVG>
+
+export default icons;
