+++ src/blocks/restaurant-menu/menu-category/save.js
@@ -1,0 +1,31 @@
+
+/**
+ * External dependencies
+ */
+import classnames from 'classnames';
+
+
+const { InnerBlocks, RichText } = wp.blockEditor
+
+const save = ( { attributes } ) => {
+
+	const {
+		menuTitle
+	} = attributes;
+
+	return (
+		<Fragment>
+			<RichText.Content
+				tagName="h1"
+				className={ classnames(
+					attributes.className ? attributes.classnames : '',
+					'kt-mc-title'
+				) }
+				value={ menuTitle }
+			/>
+			<InnerBlocks.Content />
+		</Fragment>
+	);
+};
+
+export default save;
