+++ src/blocks/restaurant-menu/save.js
@@ -1,0 +1,14 @@
+
+const {
+	InnerBlocks
+} = wp.blockEditor
+
+const save = ( { attributes } ) => {
+	return (
+		<Fragment>
+			<InnerBlocks.Content />
+		</Fragment>
+	);
+};
+
+export default save;
